This is a first step toward two simultaneous goals
1. Make it easy to implement a parallel experiment runner for Popper
2. Make Popper easier to use programatically from within another
Python application.

The major changes here are:
1. Configure the Popper loop to store information about its execution
in the Stats object. The Stats object can then be serialized or analyzed
programatically.
2. Perform all logging using python's logging framework rather than
print(). This makes it easy for other applications to tune how the
logging happens, set the log level, etc.

There should be no changes to command line Popper besides a very small
amount of extra memory usage.

For goal 1, future work will include an experiment runner in a
separate repo.

For goal 2, we probably want to add a setup.py file to build a .whl file
from Popper and publish it to PyPi.